refactor: move factories to avoid circular dep
